Move "verilog_gen" to "core_gen"
