+++ tools/backend/lsq-generator-python/core_gen/cli.py
@@ -1,6 +1,6 @@
 import argparse
-import verilog_gen.configs as configs
-import verilog_gen.codegen as code_gen
+import core_gen.configs as configs
+import core_gen.codegen as code_gen
 
 
 def parse_args():
