+++ tools/backend/lsq-generator-python/core_gen/configs.py
