+++ tools/backend/lsq-generator-python/core_gen/emitters/__init__.py
@@ -1,0 +1,8 @@
+# core_gen/emitters/__init__.py
+from core_gen.emitters.emitter import Emitter
+from core_gen.emitters.vhdl_emitter import VHDLEmitter
+from core_gen.emitters.verilog_emitter import VerilogEmitter
+
+__all__ = [
+    "Emitter", "VHDLEmitter", "VerilogEmitter"
+]
