+++ tools/backend/lsq-generator-python/core_gen/generators/README.md
