+++ tools/backend/lsq-generator-python/core_gen/generators/__init__.py
@@ -1,0 +1,10 @@
+# core_gen/generators/__init__.py
+from core_gen.generators.dispatchers import PortToQueueDispatcher, QueueToPortDispatcher
+from core_gen.generators.group_allocator import GroupAllocator
+from core_gen.generators.lsq import LSQ
+
+__all__ = [
+    "PortToQueueDispatcher", "QueueToPortDispatcher",
+    "GroupAllocator",
+    "LSQ",
+]
