+++ tools/backend/lsq-generator-python/core_gen/operators/__init__.py
@@ -1,0 +1,16 @@
+# core_gen/operators/__init__.py
+from core_gen.operators.arithmetic import WrapAdd, WrapAddConst, WrapSub
+from core_gen.operators.shifts import CyclicLeftShift
+from core_gen.operators.reduction import Reduce
+from core_gen.operators.mux import Mux1H, Mux1HROM, MuxLookUp
+from core_gen.operators.masking import CyclicPriorityMasking
+from core_gen.operators.conversions import VecToArray, BitsToOH, BitsToOHSub1, OHToBits
+
+__all__ = [
+    "WrapAdd", "WrapAddConst", "WrapSub",
+    "CyclicLeftShift",
+    "Reduce",
+    "Mux1H", "Mux1HROM", "MuxLookUp",
+    "CyclicPriorityMasking",
+    "VecToArray", "BitsToOH", "BitsToOHSub1", "OHToBits",
+]
