+++ tools/backend/lsq-generator-python/core_gen/operators/mux.py
@@ -1,8 +1,8 @@
-from verilog_gen.emitters import Emitter
-from verilog_gen.utils import *
-from verilog_gen.signals import LogicVecArray, LogicArray, Logic, LogicVec
-from verilog_gen.operators import *
-from verilog_gen.ir import Val, BinOp, WhenElse, Bit
+from core_gen.emitters import Emitter
+from core_gen.utils import *
+from core_gen.signals import LogicVecArray, LogicArray, Logic, LogicVec
+from core_gen.operators import *
+from core_gen.ir import Val, BinOp, WhenElse, Bit
 
 
 # ===----------------------------------------------------------------------===#
