+++ tools/backend/lsq-generator-python/core_gen/operators/shifts.py
@@ -1,7 +1,7 @@
-from verilog_gen.emitters import Emitter
-from verilog_gen.utils import *
-from verilog_gen.signals import *
-from verilog_gen.ir import Op, Val
+from core_gen.emitters import Emitter
+from core_gen.utils import *
+from core_gen.signals import *
+from core_gen.ir import Op, Val
 
 
 # ===----------------------------------------------------------------------===#
