+++ tools/backend/lsq-generator-python/core_gen/utils.py
