--- tools/backend/lsq-generator-python/verilog_gen/generators/__init__.py
@@ -1,10 +1,0 @@
-# verilog_gen/generators/__init__.py
-from verilog_gen.generators.dispatchers import PortToQueueDispatcher, QueueToPortDispatcher
-from verilog_gen.generators.group_allocator import GroupAllocator
-from verilog_gen.generators.lsq import LSQ
-
-__all__ = [
-    "PortToQueueDispatcher", "QueueToPortDispatcher",
-    "GroupAllocator",
-    "LSQ",
-]
