--- tools/backend/lsq-generator-python/verilog_gen/operators/__init__.py
@@ -1,16 +1,0 @@
-# verilog_gen/operators/__init__.py
-from verilog_gen.operators.arithmetic import WrapAdd, WrapAddConst, WrapSub
-from verilog_gen.operators.shifts import CyclicLeftShift
-from verilog_gen.operators.reduction import Reduce
-from verilog_gen.operators.mux import Mux1H, Mux1HROM, MuxLookUp
-from verilog_gen.operators.masking import CyclicPriorityMasking
-from verilog_gen.operators.conversions import VecToArray, BitsToOH, BitsToOHSub1, OHToBits
-
-__all__ = [
-    "WrapAdd", "WrapAddConst", "WrapSub",
-    "CyclicLeftShift",
-    "Reduce",
-    "Mux1H", "Mux1HROM", "MuxLookUp",
-    "CyclicPriorityMasking",
-    "VecToArray", "BitsToOH", "BitsToOHSub1", "OHToBits",
-]
